+++ src/lib/utils/options/options.ts
@@ -2,7 +2,6 @@ import type * as ts from "typescript";
 import { resolve } from "path";
 import { ParameterType } from "./declaration.js";
 import type { NeverIfInternal, OutputSpecification } from "../index.js";
-import { DefaultMap } from "../map.js";
 import type { Application } from "../../../index.js";
 import { insertOrderSorted, unique } from "../array.js";
 import type { Logger } from "../loggers.js";
@@ -18,6 +17,7 @@ import {
 import { addTypeDocOptions } from "./sources/index.js";
 import { getOptionsHelp } from "./help.js";
 import type { TranslationProxy } from "../../internationalization/internationalization.js";
+import { getSimilarValues } from "../general.js";
 
 /**
  * Describes an option reader that discovers user configuration and converts it to the
@@ -442,19 +442,7 @@ export class Options {
      * Discover similar option names to the given name, for use in error reporting.
      */
     getSimilarOptions(missingName: string): string[] {
-        const results = new DefaultMap<number, string[]>(() => []);
-        let lowest = Infinity;
-        for (const name of this._declarations.keys()) {
-            const distance = editDistance(missingName, name);
-            lowest = Math.min(lowest, distance);
-            results.get(distance).push(name);
-        }
-
-        // Experimenting a bit, it seems an edit distance of 3 is roughly the
-        // right metric for relevant "similar" results without showing obviously wrong suggestions
-        return results
-            .get(lowest)
-            .concat(results.get(lowest + 1), results.get(lowest + 2));
+        return getSimilarValues(this._declarations.keys(), missingName);
     }
 
     /**
@@ -491,36 +479,4 @@ export function Option<K extends keyof TypeDocOptionMap>(name: K) {
             },
         };
     };
-}
-
-// Based on https://en.wikipedia.org/wiki/Levenshtein_distance#Iterative_with_two_matrix_rows
-// Slightly modified for improved match results for options
-function editDistance(s: string, t: string): number {
-    if (s.length < t.length) return editDistance(t, s);
-
-    let v0 = Array.from({ length: t.length + 1 }, (_, i) => i);
-    let v1 = Array.from({ length: t.length + 1 }, () => 0);
-
-    for (let i = 0; i < s.length; i++) {
-        v1[0] = i + 1;
-
-        for (let j = 0; j < s.length; j++) {
-            const deletionCost = v0[j + 1] + 1;
-            const insertionCost = v1[j] + 1;
-            let substitutionCost: number;
-            if (s[i] === t[j]) {
-                substitutionCost = v0[j];
-            } else if (s[i]?.toUpperCase() === t[j]?.toUpperCase()) {
-                substitutionCost = v0[j] + 1;
-            } else {
-                substitutionCost = v0[j] + 3;
-            }
-
-            v1[j + 1] = Math.min(deletionCost, insertionCost, substitutionCost);
-        }
-
-        [v0, v1] = [v1, v0];
-    }
-
-    return v0[t.length];
 }